frontend/gamma/tests/tests/Clipperz/Crypto/RSA.html

<!--

Copyright 2008-2015 Clipperz Srl

This file is part of Clipperz, the online password manager.
For further information about its features and functionalities please
refer to http://www.clipperz.com.

* Clipperz is free software: you can redistribute it and/or modify it 
  under the terms of the GNU Affero General Public License as published
  by the Free Software Foundation, either version 3 of the License, or 
  (at your option) any later version.

* Clipperz is distributed in the hope that it will be useful, but 
  WITHOUT ANY WARRANTY; without even the implied warranty of 
  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
  See the GNU Affero General Public License for more details.

* You should have received a copy of the GNU Affero General Public
  License along with Clipperz. If not, see http://www.gnu.org/licenses/.

-->

<html>
<head>
	<script type="text/javascript" src="../../../../js/MochiKit/MochiKit.js"></script>
    <script type="text/javascript" src="../../../SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" type="text/css" href="../../../SimpleTest/test.css">

	<script type='text/javascript' src='../../../../js/Clipperz/Base.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/DOM.js'></script>

	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/BigInt.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/Base.js'></script>
	<script type='text/javascript' src='../../../../js/Clipperz/Crypto/RSA.js'></script>

</head>
<body>
<pre id="test">
<script type="text/javascript">
try {
	var	rsaPublicKey;
	var	randomValue;
	var	publicEncryptedValue;
	var privateEncryptedValue;

	var	startTime;
	var	endTime;


startTime=new Date();
	rsaPublicKey = Clipperz.Crypto.RSA.generatePublicKey(512);

endTime=new Date();
is(true, true, 'Time to generate the RSA public key (size 512 bit): '+(endTime.getTime()-startTime.getTime())/1000.0);

	randomValue = Clipperz.Crypto.Base.generateRandomSeed();
	publicEncryptedValue = Clipperz.Crypto.RSA.encryptUsingPublicKey(rsaPublicKey, randomValue);
	privateEncryptedValue = Clipperz.Crypto.RSA.encryptUsingPrivateKey(rsaPublicKey, randomValue);
	
	is(publicEncryptedValue == privateEncryptedValue, false);
	is(Clipperz.Crypto.RSA.decryptUsingPrivateKey(rsaPublicKey, publicEncryptedValue), randomValue);
	is(Clipperz.Crypto.RSA.decryptUsingPublicKey(rsaPublicKey, privateEncryptedValue), randomValue);

	//-------------------------------------------------------------------------
} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );
}

</script>
</pre>
</body>
</html>
